(GH-31) (GH-42) added searchPaths and excludes
a new settings page was introduced: "Search paths".
This page contains two settings: The search paths to start
the search for Cake files and excludes for paths that
should not be searched.
Default search paths is ["."], which correlates to the "old"
functionality.
Default exclude is [".*/tools/.*"] to exclude the tools
folder and thereby excluding Cake files that were
added from NuGet packages.