(GH-113) added OperatingSystem and limited
StackTrace to 5000 chars.
For long StackTraces a "please add StackTrace here, manually" hint
appears which is not ideal but gets the job done.

For the future we should either add some gh-api access (Octokit)
which would mean the user needs to provide OAuth tokens or something..
Alternatively we could host a small web service somewhere to create an
issue in GitHub anonymously (rather: Under some bot-account).